requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
databases
asyncpg
python-dotenv
sqlalchemy
psycopg2-binary
